Fix task
